Compute timestamp offset
The VDIF timestamp is only accurate to within 8 microseconds (data
length of one packet), whereas the B-engine timestamp reports time
to within ~0.25ns. The current implementation packs output VDIF data
starting with the first sample in the first B-engine frame so that
there may be an up to 8 microseconds offset in the timestamp applied.
To account for this the offset is calculated and reported in the log
to stdout.

+++ software/aphids/beng_reform.c
@@ -174,8 +174,8 @@ void beng_reform_headers(vdif_in_packet_t *pkts, int npkts) {
 	}
 }
 
-float beng_timestamp_clk_to_float(beng_timestamp_t *t) {
-	return (float)(t->clk) / (float)SWARM_CLK_PER_SEC;
+double beng_timestamp_clk_to_float(beng_timestamp_t *t) {
+	return (double)(t->clk) / (double)SWARM_CLK_PER_SEC;
 }
 
 void beng_timestamp_increment(beng_timestamp_t *t) {
